requirements.txt
<<<<<<< HEAD
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
=======
google-api-python-client==2.99.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
>>>>>>> 00a560fbf5feea7c75937daaf4ed9253f986b2b3
